requirements.txt
python-kasa>=0.5.0 